requirements.txt
gradio>=4.0.0
openai>=1.0.0
requests>=2.31.0
pillow>=10.0.0
gtts>=2.4.0
pygame>=2.5.0
python-dotenv>=1.0.0
transformers>=4.35.0
torch>=2.1.0
diffusers>=0.24.0
accelerate>=0.24.0